requirements.txt
pillow
nibabel
imageio
requests
tqdm
fastapi
uvicorn
python-dotenv
supabase
pydicom
numpy
torch
opencv-python
scikit-image
python-dotenv
python-multipart